Move isStreamCompressed into copy/compression.go
copy/copy.go is becoming too big, shrink it at least a bit by separating
the most clearly independent part of the code.
